+++ frontend/src/App.js
@@ -11,8 +11,10 @@ function App() {
   return (
     <body>
       <div class="topnav">
-        <Link to="/temas"><a class="navbarBoton">Cursos</a></Link>
+        <Link to="/temas"><a class="navbarBoton">Temas</a></Link>
         <Link to="/alumnos"><a class="navbarBoton">Alumnos</a></Link>
+        <Link to="/crearTema"><a class="navbarBoton">Crear Tema</a></Link>
+        <Link to="/crearCurso/0"><a class="navbarBoton">Crear Curso</a></Link>
       </div>
       <Routes>
           <Route path="/temas/*" element={<Temas />} />